pyproject.toml
[tool.poetry]
name = "FloofyBot"
version = "0.1.0"
description = "A discord bot made with naff."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
orjson = "^3.6.7"
naff = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
